fix(consul): delete document
- rewritten with `documentAttributes` props useState in
  `Document Selector` due to deletion error
- added `useEffect` to run onChange function according
  to `documentsAttributes` changes in `DocumentSelector`
- change the type of `defaultProps` added for documents
  in `NewProposal.js` due to an error when trying to edit a
  proposal that contains a document, to array
- write error message generators in one file
- add sperator to `newItems` in `deleteArrayItem` helper
  because index in main array is deleted

SVA-596
