+++ src/helpers/consul/index.js
@@ -1,5 +1,4 @@
-export * from './documentErrorMessageGenerator';
+export * from './dataErrorMessageGenerator';
 export * from './filterHelper';
 export * from './homeData';
-export * from './imageErrorMessageGenerator';
 export * from './sortingHelper';
